ecommerce_project/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-auditlog==2.1.1
django-filter==22.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
idna==3.4
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==2.0.3
